+++ spec/dummy/client/entries/app-kitten.js
@@ -206,6 +206,7 @@ import { KarlMenu,
 import { KarlBrowsing } from 'app/browsing/browsing'
 import { KarlSeparator } from 'kitten/karl/layout/separator'
 import { KarlImageCropper } from 'kitten/karl/images/image-cropper'
+import { KarlModal } from 'kitten/karl/modals/modal'
 
 ReactOnRails.register({
   // Box
@@ -405,4 +406,5 @@ ReactOnRails.register({
   KarlGithubRibbon,
   KarlSeparator,
   KarlImageCropper,
+  KarlModal,
 })